Revisio USB device MSC

+++ v2/Include/Controllers/USBDevice/ST/st_usbd_def.h
@@ -315,8 +315,8 @@ typedef struct
 /* USB Device handle structure */
 struct USBD_HandleTypeDef {
 	uint8_t                 id;
-	uint32_t                dev_config;
-	uint32_t                dev_default_config;
+	uint8_t                 dev_config;
+	uint8_t                 dev_default_config;
 	uint32_t                dev_config_status;
 	USBD_SpeedTypeDef       dev_speed;
 	USBD_EndpointTypeDef    ep_in[16];
